explicitly set updated_at attribute for playlists to avoid  non-deterministic sort of radio.playlists
- basically the issue is being caused because when rspec is run on CI/CD all the records sometimes have created_at and updated_at set to exactly "2015-01-01 00:00:00"
- this happened because a random test, spec/models/scheduled_show_spec.rb, called `Timecop.freeze Time.local(2015)`, but never called `Timecop.return`.  this means that every subsequent spec ran with time frozen at exactly new year's 2015, this may be responsible for multiple intermittent spec failures that depend on timestamps
- in the case of playlist_spec.rb, when the existing default playlist is destroyed, the model first applies a default_scope of `order(updated_at: :desc)` that sorts by updated date, and then chooses the playlist that has the latest updated_at to become the new default thru `radio.playlists.where.not(id: id).first.id`.  if all the playlists have the same timestamp, then the ActiveRecord query will return a random order each time, thus causing the intermittent failure
- I've done a few things:
(1) I've added a `Timecop.return` to the spec that called `Timecop.freeze` so that time is not frozen for all subsequent specs
(2) renamed the callback method in the Playlist model to be more explicit about what it's doing to avoid confusion - it's not just setting any old playlist as the new default; rather , it's setting the most recently updated, as defined by the updated_at timestamp, as default.
(3) since the spec is really about picking whether playlist_1 or playlist_2 is the new default, I explicitly set their updated_at to be 1 day apart.  so that it is clear that it is the later updated_at that is the deciding factor for which playlist becomes default